🙈 Allow userId query parameter (#442)
There is a bug in most virtru-sdk for javascript clients that currently sends the userId in some KAS requests, instead of just ones to the virtru auth provider. I've sent a PR to fix this in the client, but we need to work around existing clients in the wild to prevent them from getting 400s.